resources: reorganize folder
