+++ invenio_drafts_resources/resources/drafts/schemas_links.py
