+++ invenio_drafts_resources/resources/records/__init__.py
@@ -11,8 +11,10 @@
 
 from .config import RecordResourceConfig
 from .resource import RecordResource
+from .schemas_links import RecordLinksSchema
 
 __all__ = (
+    "RecordLinksSchema",
     "RecordResource",
     "RecordResourceConfig",
 )
