+++ setup.py
@@ -59,7 +59,7 @@ setup_requires = [
 
 install_requires = [
     "invenio_pidrelations>=v1.0.0a7,<2.0.0",
-    "invenio-records-resources>=0.6.3,<0.7.0",
+    "invenio-records-resources>=0.6.4,<0.7.0",
 ]
 
 packages = find_packages()
